# Metric SDK - remove de-duplicating attributes for perf gain