--- a/opentelemetry-sdk/src/metrics/mod.rs
+++ b/opentelemetry-sdk/src/metrics/mod.rs
@@ -160,6 +160,66 @@ mod tests {
     // "multi_thread" tokio flavor must be used else flush won't
     // be able to make progress!
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
+    async fn counter_aggregation_duplicate_attributes() {
+        // Run this test with stdout enabled to see output.
+        // cargo test counter_aggregation_duplicate_attributes --features=metrics,testing
+
+        // Arrange
+        let exporter = InMemoryMetricsExporter::default();
+        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
+        let meter_provider = SdkMeterProvider::builder().with_reader(reader).build();
+
+        // Act
+        let meter = meter_provider.meter("test");
+        let counter = meter
+            .u64_counter("my_counter")
+            .with_unit(Unit::new("my_unit"))
+            .init();
+        counter.add(1, &[KeyValue::new("key1", "value1"),KeyValue::new("key1", "value2")]);
+        counter.add(1, &[KeyValue::new("key1", "value1"),KeyValue::new("key1", "value2")]);
+
+        meter_provider.force_flush().unwrap();
+
+        // Assert
+        let resource_metrics = exporter
+            .get_finished_metrics()
+            .expect("metrics are expected to be exported.");
+        assert!(!resource_metrics.is_empty());
+        let metric = &resource_metrics[0].scope_metrics[0].metrics[0];
+        assert_eq!(metric.name, "my_counter");
+        assert_eq!(metric.unit.as_str(), "my_unit");
+        let sum = metric
+            .data
+            .as_any()
+            .downcast_ref::<data::Sum<u64>>()
+            .expect("Sum aggregation expected for Counter instruments by default");
+
+        // Expecting 1 time-series.
+        assert_eq!(sum.data_points.len(), 1);
+
+        // find and validate key1=value1,key1=value2 datapoint        
+        let data_point = &sum.data_points[0];
+        assert_eq!(
+            data_point
+                .value,
+            2
+        );
+        assert_eq!(data_point.attributes.len(), 2, "Should have 2 attributes as sdk is not deduplicating attributes");
+        let key_value1_found = data_point
+            .attributes
+            .iter()
+            .any(|(k, v)| k.as_str() == "key1" && v.as_str() == "value1");
+        let key_value2_found = data_point
+            .attributes
+            .iter()
+            .any(|(k, v)| k.as_str() == "key1" && v.as_str() == "value2");
+
+        assert_eq!(key_value1_found && key_value2_found, true, "Should have found both key1=value1 and key1=value2 attributes as does not dedup");
+    }
+
+    // "multi_thread" tokio flavor must be used else flush won't
+    // be able to make progress!
+    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     async fn counter_duplicate_instrument_merge() {
         // Arrange
         let exporter = InMemoryMetricsExporter::default();