@hdost 2023-11-23
So going back to our conversation from before the gain here is not the 7x we saw. 
I guess my interpretation was that we'd look at bounded instruments and/or allowing users to pass in prebuilt AttributeSets. 
If we're going to just break it this seems like we should put it behind a feature flag which was also something we discussed.