@hdost 2023-11-25
TL;DR I think we should add the feature flag before merging this. 

> > So going back to our conversation from before the gain here is not the 7x we saw. This is closer to a 6-8% improvement based on the stress test.
> 
> 
> 
> There is no 7x gain with this PR. I am trying to understanding why did you expect 7x here? Apologies if I missed something obvious..
> 

My point was maybe not well made.
https://github.com/open-telemetry/opentelemetry-rust/pull/1379 we saw significant improvement on the order of 6-7x improvement (from what I understand and doesn't break any previous functionality. 
This PR breaks backwards compatibility and is not as significant improvement. 


> > we'd look at bounded instruments and/or allowing users to pass in prebuilt AttributeSets
> 
> Yes absolutely! That is a separate improvement, helping users who has key/values known ahead. This PR is not attempting that improvement.
> 

I think we'd be better off focusing on providing those to provide performance. And if we're going to break functionality I really think we need to be more thoughtful about it. Yes we want performance, but if we are constantly breaking things people have little incentive to move forward and with our latest release I don't think we've done a good job of providing a path forward. 

This is a little wider point but it's relevant. I don't want to be the barrier to progress, but I also don't want no one to use this because they experience frustration through upgrades. 

> 
> > If we're going to just break it this seems like we should put it behind a feature flag
> 
> 
> 
> If users have duplicate keys in their attributes, that is already an indication of something wrong with their instrumentation. I don't think its fair to ask every user to pay for the perf penality to dedup! I am totally open to bringing de-dup as an opt-in feature, which is why I has opened [this issue to gather feedback](https://github.com/open-telemetry/opentelemetry-rust/issues/1300). 
> 

On the point above I think that we should have a feature flag on this to allow people to work with the old behavior. Feedback is great but we don't give people the option to try both ways as they would have to use an older version which might miss other internal improvements.