@hdost 2023-11-27
> TL;DR I think we should add the feature flag before merging this.

> I strongly believe that the **default behavior should prioritize efficiency**, and I fully support the idea of a feature flag or opt-in approach for the less efficient route involving deduplication, rather than defaulting to it.

I guess we're on the same page that there should be a feature flag.

As I suggested it could be something like: `metrics-enable-deduplication`