@KallDrexx 2023-11-27
Maybe it's prudent to have concrete idea of what the ramifications of not-deduplicating is?   From my limited knowledge, there are two ramifications

1. You may end up with duplicated time series (e.g. `{color: red, shape: circle, color:red}` vs `{color:red, shape: cicle}` would be different time series). 
2. Some exporters may fail (e.g. prometheus non-otlp).

There's always a balance between performance and creating the pit of success for users, and it's hard for me to give much of an opinion without how accidentally violating the duplicate key rules in this case actually works.

Fwiw, I don't have a problem enabling this to be opt-in as a "I am aware of how I build my attributes and explicitly know that providing duplicate attributes may cause bad data".  I'm a big proponent of pit of success, and a feature flag to enable better performance without code changes doesn't seem too erroneous.