@KallDrexx 2023-11-27
I guess my primary concerns is with `{color: red, shape: circle, color:red}` vs `{color:red, shape: cicle}`.

If we don't dedup, then these are stored in the `ValueMap` as two different attribute sets, which means a counter against one won't be applied to the other.  This means you could have two different values in both attribute sets.

I assume the backends will automatically aggregate them into a single time series in that case?