@cijothomas 2023-11-28
> I assume the backends will automatically aggregate them into a single time series in that case?

Again, this is backend specific.