@cijothomas 2023-12-07
Closing for now, as we are exploring other options, and will come back to this after https://github.com/open-telemetry/opentelemetry-rust/pull/1421